:lipstick: fix narrow-width crowding + oversized readouts in 5 frames
- volume-profile: 2x2 stat header (was 4-across) so full price levels fit
  without clipping; tiles shrinkable + truncate as a safety net.
- national-debt: total/delta row wraps at narrow width so the delta and
  "since <date>" no longer clip/mangle.
- yield-risk-pie: center total metric-lg -> metric-md + width clamp so it fits
  inside the donut hole instead of clipping against the ring.
- returns-projector: compact huge returns ("+517T%") past 5 digits and guard
  non-finite ending/gain, instead of a 15-digit percentage.
- funding-heatmap: numeric "M/D" date labels; HeatmapChart now thins column
  labels (shows every Nth in full) when cells are too narrow to hold one,
  rather than truncating every column to an indistinct "7…".

HeatmapChart thinning only triggers on narrow cells; wide heatmaps
(fx-cross, coin-momentum, funding-venue) verified unchanged. Verified in
Storybook; typecheck + 913 tests green.

+++ packages/frames/src/funding-heatmap.tsx
@@ -47,8 +47,10 @@ function FundingHeatmap({ config }: { config: z.output<typeof schema> }) {
       for (const [bucket, { sum, n }] of buckets) {
         const avg = sum / n;
         const date = new Date(bucket);
+        // Numeric "M/D" ("7/17") stays legible in a narrow daily column;
+        // "Jul 17" gets truncated to an indistinct "J…" when space is tight.
         const column = date.toLocaleDateString("en-US", {
-          month: "short",
+          month: "numeric",
           day: "numeric",
         });
         out.push({
